blog-backend/requirements.txt
# System libraries
envparse==0.2.0
psycopg2-binary==2.9.2

# Django related libraries
Django==3.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
dj-rest-auth==2.2.4
django-cors-headers==3.11.0
django-extensions==3.1.5
django-admin-list-filter-dropdown==1.0.3

# Other
requests==2.27.1
gunicorn==20.1.0
requests-ntlm==1.1.0

# Datetime related (without version, always fetch latest)
python-dateutil
pytz

# Development
coverage==6.2
factory-boy==3.2.1

argh==0.26.2
flake8==3.9.2
flake8-2020==1.6.0
flake8-broken-line==0.3.0
flake8-bugbear==21.4.3
flake8-builtins==1.5.3
flake8-commas==2.0.0
flake8-comprehensions==3.5.0
flake8-debugger==4.0.0
flake8-django==1.1.2
flake8-eradicate==1.1.0
flake8-expression-complexity==0.0.9
flake8-fixme==1.1.1
flake8-isort==4.0.0
flake8-logging-format==0.6.0
flake8-multiline-containers==0.0.18
flake8-mutable==1.2.0
flake8-pie==0.14.0
flake8-polyfill==1.0.2
flake8-print==4.0.0
flake8-printf-formatting==1.1.2
flake8-quotes==3.2.0
flake8-simplify==0.14.1
flake8-use-fstring==1.1
fuzzywuzzy==0.18.0
pep8-naming==0.12.0
pylint==2.9.5
pyaml==20.4.0
pylint==2.9.5
pylint-fail-under==0.3.0
pyflakes==2.3.1
pytest==6.2.4
pytest-asyncio==0.15.1
PyYAML==5.4.1

watchdog==2.1.7
